chore: rework main so names make sense
