+++ lib/controllers/backpack.js
@@ -10,7 +10,7 @@ const {
   GAMMA_DEFAULT,
   FIRMATA_7BIT_MASK
 } = require('../constants');
-const { Pixel } = require('../index')
+const { Pixel } = require('../pixel');
 const { create_gamma_table } = require('../utils');
 
 const IC2Backpack = {
