+++ lib/controllers/firmata.js
@@ -14,7 +14,7 @@ const {
   PIXEL_COMMAND
 } = require('../constants');
 const { create_gamma_table } = require('../utils');
-const { Pixel } = require('../index')
+const { Pixel } = require('../pixel');
 
 const Firmata = {
   initialize: {
