+++ lib/strip.js
@@ -1,6 +1,6 @@
 const events = require('events');
 const util = require('util');
-const { ColorString } = require('./index')
+const { ColorString } = require('./pixel');
 const {
   SHIFT_FORWARD
 } = require('./constants');
