+++ test/i2cbackpack.js
@@ -6,7 +6,7 @@ const MockSerialPort = mocks.SerialPort;
 const sinon = require('sinon');
 
 const five = require('johnny-five');
-const pixel = require('../lib/pixel.js');
+const pixel = require('../lib/index.js');
 
 const Board = five.Board;
 
@@ -79,9 +79,9 @@ exports['Strip - I2C'] = {
         if (err) {
           if (err.name == 'NoIOError') {
             return true;
-          } 
+          }
           return false;
-        } 
+        }
         return false;
       },
       'If IO is not present an error should be thrown'
